test(ch): strict-scan differential vs real ClickHouse
Add a Layer 6e harness that executes the matrix-shaped spec golden SQL
corpus against a REAL ClickHouse (testcontainers-go) through the
production cursor's strict positional scan, failing on any type-coercion
error. This catches the documented prod incident class the chdb-tagged
goldens are structurally blind to: chDB leniently coerces result-column
types into the Go Scan destinations (e.g. a UInt8/UInt64 Value column
into *float64), while prod clickhouse-go/v2 strict-scans and 502s. The
chdb cells only ever observe the COERCED value, so they can never catch
an emit-type bug; this lane re-runs the same emitted SQL (native Map /
DateTime64 / Float64 columns, NOT the chDB toJSONString Map-wrap that
masks the divergence) through the exact production cursor scan.

- test/spec/strictscan_integration_test.go (integration tag): walks the
  round-trip fixtures under test/spec/{promql,logql,traceql}, probes each
  emitted SQL's result column shape, and strict-scans only the
  matrix-shaped ones (MetricName, Attributes, TimeUnix, Value [,Metadata])
  through chclient.QueryCursor. Mirrors the existing chclient/schema
  testcontainers integration pattern.
- roundtrip_prep.go: hoist the build-tag-free seed/SQL prep pipeline
  (substituteNow64, splitStatements, promoteCreateTable,
  backfillResourceAttributes + string helpers) out of the chdb-tagged
  runner into one shared, untagged file so the chDB runner, the perf
  profiler, and the strict-scan differential share one copy.
- splitStatements is now comment-aware: a stray apostrophe inside a `--`
  seed comment (e.g. "the runner's `SELECT *`") previously flipped the
  in-string state machine and glued CREATE+INSERT into one statement;
  chdb-go tolerated the multi-statement, clickhouse-go's native Exec
  rejects it (code 62). Fix is a strict improvement; full chDB roundtrip
  corpus (promql+logql+traceql) stays green.
- strict-scan.yml: informational CI lane on PR + push + nightly +
  dispatch (NOT a required gate). Justfile `strict-scan-test` recipe.
- docs/test-strategy.md: Layer 6e row + CI-gate inventory row.

Result: PASSES clean over 334 matrix-shaped fixtures against ClickHouse
25.8 — confirming the Value-typed emitters (metric/range/vector-join)
correctly toFloat64-wrap today. The harness's catch-power is proven: the
pre-scoping run surfaced 13 real Map->*string / UInt8->*string column
mismatches plus the splitStatements latent bug.
